added header count products
