+++ src/components/layout/header/header.module.scss
@@ -9,3 +9,15 @@
   display: flex;
   column-gap: 28px;
 }
+
+.item {
+  position: relative;
+}
+
+.decor {
+  padding: 6px 12px;
+  background-color: #f5f5f5;
+  border-radius: 50%;
+  color: #2b6cb0;
+  margin-left: 4px;
+}
