+++ src/shared/constants/navigation.ts
@@ -4,19 +4,20 @@ const NAVIGATION = [
     alias: '/',
     name: "Список товаров",
   },
+  {
+    id: 2,
+    alias: '/cart',
+    name: "Корзина",
+    showCart: true
+  },
 ]
 
 export const NAVIGATION_ACTIVE_USER = [
   ...NAVIGATION,
   {
-    id: 2,
+    id: 3,
     alias: '/create-product',
     name: "Создать продукт",
-  },
-  {
-    id: 3,
-    alias: '/cart',
-    name: "Корзина",
   },
   {
     id: 4,
@@ -25,27 +26,24 @@ export const NAVIGATION_ACTIVE_USER = [
     exit: true
   },
 ] as {
-  id: number,
-  alias: string,
-  name: string,
-  exit?: true
+  id: number;
+  alias: string;
+  name: string;
+  exit?: true;
+  showCart?: true;
 }[]
 
 export const NAVIGATION_INACTIVE_USER = [
   ...NAVIGATION,
-  {
-    id: 2,
-    alias: '/cart',
-    name: "Корзина",
-  },
   {
     id: 3,
     alias: '/authentication',
     name: "Вход",
   },
 ] as {
-  id: number,
-  alias: string,
-  name: string,
-  exit?: true
+  id: number;
+  alias: string;
+  name: string;
+  exit?: true;
+  showCart?: true;
 }[]
